Fixed taskWidget color bug, added the correct data display about a task in the task manager screen, created task_creation_view_model.dart and moved the AddTask,selectColor and selectDate functions from task_creation_screen_view.dart to the created file. Done with bountx :)
