+++ pwi_project/lib/view/task_creation_screen_view.dart
@@ -1,73 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_colorpicker/flutter_colorpicker.dart';
 import 'package:provider/provider.dart';
-
 import '../model/task.dart';
+import '../view_model/task_creation_view_model.dart';
 import '../view_model/task_view_model.dart';
 
-class TaskCreationViewModel extends ChangeNotifier {
-  TextEditingController titleController = TextEditingController();
-  TextEditingController descriptionController = TextEditingController();
-  DateTime selectedDate = DateTime.now();
-  Color selectedColor = Colors.amber;
 
-      Future<void> _selectDate(BuildContext context) async {
-        final DateTime? picked = await showDatePicker(
-          context: context,
-          initialDate: selectedDate,
-          firstDate: DateTime.now(),
-          lastDate: DateTime(2101),
-        );
-        if (picked != null && picked != selectedDate) {
-          selectedDate = picked;
-          notifyListeners();
-        }
-  }
-
-  void _selectColor(BuildContext context) {
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Pick a color!'),
-          content: SingleChildScrollView(
-            child: ColorPicker(
-              pickerColor: selectedColor,
-              onColorChanged: (Color color) {
-                selectedColor = color;
-                notifyListeners();
-              },
-              showLabel: true,
-              pickerAreaHeightPercent: 0.8,
-            ),
-          ),
-          actions: <Widget>[
-            TextButton(
-              child: const Text('Got it'),
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  void addTask(BuildContext context) {
-    Provider.of<TaskList>(context, listen: false).addTask(
-      Task(
-        titleController.text,
-        descriptionController.text,
-        selectedDate,
-        selectedColor,
-        false,
-      ),
-    );
-    Navigator.pop(context);
-    notifyListeners();
-  }
-}
 
 class TaskCreationWidget extends StatelessWidget {
   @override
@@ -106,13 +44,13 @@ class TaskCreationWidget extends StatelessWidget {
                 SizedBox(height: 20),
                 Text('Finish Date: ${model.selectedDate.toLocal()}'),
                 ElevatedButton(
-                  onPressed: () => model._selectDate(context),
+                  onPressed: () => model.selectDate(context),
                   child: Text('Select Finish Date'),
                 ),
                 SizedBox(height: 20),
                 Text('Background Color:'),
                 ElevatedButton(
-                  onPressed: () => model._selectColor(context),
+                  onPressed: () => model.selectColor(context),
                   style: ElevatedButton.styleFrom(
                       backgroundColor: model.selectedColor),
                   child: Text('Select Color'),
