+++ pwi_project/lib/view/task_manager_screen_view.dart
@@ -48,10 +48,10 @@ class TaskManagerScreen extends StatelessWidget {
               return Padding(
                 padding: const EdgeInsets.fromLTRB(8.0, 10.0, 8.0, 0.0),
                 child: TaskWidget(task: Task(
-                  'Name',
-                  'Zawartość',
-                  DateTime.now(),
-                  Colors.pink,
+                  taskList.tasks[index].name,
+                  taskList.tasks[index].description,
+                  taskList.tasks[index].day,
+                  taskList.tasks[index].background,
                   false
                 ),
                 ),
